requirements.txt
annotated-types==0.7.0
anyio==4.8.0
assemblyai==0.37.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
gtts==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.0.2
pyaudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygame==2.6.1
python-dotenv==1.0.1
requests==2.32.3
scipy==1.13.1
sniffio==1.3.1
sounddevice==0.5.1
srt==3.5.3
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
vosk==0.3.45
websockets==14.2
